backend/requirements.txt
fastapi[all]==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
anthropic==0.50.0
litellm==1.67.2
arxiv==2.2.0